src/spotify-mcp-server/pyproject.toml
[project]
name = "spotify-mcp-server"
version = "0.1.0"
description = "MCP server exposing Spotify API via Spotipy"
authors = ["Benoit Moussaud <your.email@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "spotipy",
    "fastmcp>=2.12.4",
    "azure-monitor-opentelemetry",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-httpx",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-starlette"
]

[tool.setuptools.packages.find]
where = ["."]
